Improvements to `using Static` time (#55)
For Julia v1.6 this changes @time using Static from 0.054736 to 0.043993.
For Julia v1.7 this changes @time using Static from 0.048681 to 0.038870.

A lot of the improvement here is because redundant method definitions identified in the StaticInt PR to Base (e.g., Base.:(*)(::Zero, ::Zero)). As discussed in that PR, these are already known at compile time so explicitly defining them doesn't improve performance.

I spent a pretty good amount of time with SnoopCompile and identified and simplified some method influence compile time. I tried generating a precompile file but it made start up time worse, so I'll have to look into that more later. This might be as good of an improvement as we can get while we still have invalidations in Base we can't resolve.